UPDATE TO V0.22.0

+++ src-plugins/logpipe-output-tcp.c
@@ -356,10 +356,10 @@ _GOTO_RETRY_SEND :
 }
 
 funcWriteOutputPlugin WriteOutputPlugin ;
-int WriteOutputPlugin( struct LogpipeEnv *p_env , struct LogpipeOutputPlugin *p_logpipe_output_plugin , void *p_context , uint32_t file_offset , uint32_t file_line , uint32_t block_len , char *block_buf )
+int WriteOutputPlugin( struct LogpipeEnv *p_env , struct LogpipeOutputPlugin *p_logpipe_output_plugin , void *p_context , uint64_t file_offset , uint64_t file_line , uint64_t block_len , char *block_buf )
 {
 	struct OutputPluginContext	*p_plugin_ctx = (struct OutputPluginContext *)p_context ;
-	uint32_t			block_len_htonl ;
+	uint64_t			block_len_htonl ;
 	int				len ;
 	
 	int				nret = 0 ;
@@ -418,7 +418,7 @@ int AfterWriteOutputPlugin( struct LogpipeEnv *p_env , struct LogpipeOutputPlugi
 {
 	struct OutputPluginContext	*p_plugin_ctx = (struct OutputPluginContext *)p_context ;
 	
-	uint32_t			block_len_htonl ;
+	uint64_t			block_len_htonl ;
 	
 	struct timeval			tv_diff_send_filename ;
 	struct timeval			tv_diff_send_block_len ;
